Refactored DataMatcher into RulesDataMatcher.
