+++ src/Plugin/RulesDataMatcherPluginManager.php
@@ -12,6 +12,8 @@ use Drupal\Core\Plugin\DefaultPluginManager;
 
 /**
  * Plugin manager for all Rules' DataMatchers.
+ *
+ * @see \Drupal\rules\Matcher\MatcherInterface
  */
 class RulesDataMatcherPluginManager extends DefaultPluginManager {
 
@@ -19,9 +21,8 @@ class RulesDataMatcherPluginManager extends DefaultPluginManager {
    * {@inheritdoc}
    */
   public function __construct(\Traversable $namespaces, ModuleHandlerInterface $module_handler, $plugin_definition_annotation_name = 'Drupal\rules\Annotation\RulesDataMatcher') {
-    $this->alterInfo('rules_datamatcher');
-
-    parent::__construct('Plugin/DataMatcher', $namespaces, $module_handler, 'Drupal\rules\Matcher\MatcherInterface', $plugin_definition_annotation_name);
+    $this->alterInfo('rules_data_matcher');
+    parent::__construct('Plugin/RulesDataMatcher', $namespaces, $module_handler, 'Drupal\rules\Matcher\MatcherInterface', $plugin_definition_annotation_name);
   }
 
 }
